Add things in RapidsConf
