+++ integration_tests/src/assembly/bin.xml
@@ -51,7 +51,7 @@
             <outputDirectory>integration_tests</outputDirectory>
         </file>
         <file>
-            <source>${project.build.directory}/spark-rapids-default-configs.json</source>
+            <source>${spark.rapids.source.basedir}/dist/target/spark-rapids-default-configs.json</source>
             <outputDirectory>integration_tests</outputDirectory>
         </file>
     </files>
